Rename file to CamelCase(TypeDescriptor)
