+++ packages/core/src/Engine/Engine.ts
@@ -3,7 +3,7 @@ import timecodes from 'node-timecodes'
 
 import { Clip, Effect, Project } from '../Entity'
 import EffectPluginBase from '../PluginSupport/PostEffectBase'
-import { ParameterValueTypes } from '../PluginSupport/type-descriptor'
+import { ParameterValueTypes } from '../PluginSupport/TypeDescriptor'
 
 import { IRenderingStreamObserver, RenderingStatus } from './IRenderingStreamObserver'
 import { IRenderer } from './Renderer/RendererBase'
