+++ packages/core/src/Engine/ParametersTable.ts
@@ -3,7 +3,7 @@ import _ from 'lodash'
 import { EffectRenderContext } from '..'
 import { Clip, Keyframe } from '../Entity'
 import { UserCodeException } from '../Exceptions/UserCodeException'
-import { ParameterValueTypes, TypeDescriptor } from '../PluginSupport/type-descriptor'
+import { ParameterValueTypes, TypeDescriptor } from '../PluginSupport/TypeDescriptor'
 import { AssetPointer, ColorRGB, ColorRGBA, Expression } from '../Values'
 import AssetProxy from './AssetProxy'
 import { compileTypeScript } from './ExpressionSupport/ExpressionCompiler'
