+++ packages/core/src/Engine/Renderer/Audio/Audio.ts
@@ -2,8 +2,8 @@ import fs from 'fs'
 import _ from 'lodash'
 
 import { resampling } from '../../../helper/Audio'
-import Type from '../../../PluginSupport/type-descriptor'
-import { TypeDescriptor } from '../../../PluginSupport/type-descriptor'
+import Type from '../../../PluginSupport/TypeDescriptor'
+import { TypeDescriptor } from '../../../PluginSupport/TypeDescriptor'
 import { IRenderer } from '../RendererBase'
 
 import { Asset } from '../../../Entity'
