+++ packages/core/src/Engine/Renderer/Image/Image.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import Type from '../../../PluginSupport/type-descriptor'
+import Type from '../../../PluginSupport/TypeDescriptor'
 import { IRenderer } from '../RendererBase'
 
 import { Asset } from '../../../Entity'
