+++ packages/core/src/Engine/Renderer/RendererBase.ts
@@ -1,4 +1,4 @@
-import { TypeDescriptor } from '../../PluginSupport/type-descriptor'
+import { TypeDescriptor } from '../../PluginSupport/TypeDescriptor'
 import { ClipPreRenderContext } from '../RenderContext/ClipPreRenderContext'
 import { ClipRenderContext } from '../RenderContext/ClipRenderContext'
