+++ packages/core/src/Engine/Renderer/Video/Video.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 
-import Type from '../../../PluginSupport/type-descriptor'
-import { TypeDescriptor } from '../../../PluginSupport/type-descriptor'
+import Type from '../../../PluginSupport/TypeDescriptor'
+import { TypeDescriptor } from '../../../PluginSupport/TypeDescriptor'
 import { IRenderer } from '../RendererBase'
 
 import { Asset } from '../../../Entity'
