+++ packages/core/src/Engine/Task/EffectRenderTask.ts
@@ -3,7 +3,7 @@ import _ from 'lodash'
 import { Clip, Effect } from '../../Entity'
 import { EffectPluginMissingException } from '../../Exceptions'
 import EffectPluginBase from '../../PluginSupport/PostEffectBase'
-import { TypeDescriptor } from '../../PluginSupport/type-descriptor'
+import { TypeDescriptor } from '../../PluginSupport/TypeDescriptor'
 import DependencyResolver from '../DependencyResolver'
 import * as ExpressionContext from '../ExpressionSupport/ExpressionContext'
 import { ReferenceableEffectsParams } from '../ExpressionSupport/ExpressionContext'
