+++ packages/core/src/PluginSupport/PostEffectBase.ts
@@ -1,6 +1,6 @@
 import { EffectPreRenderContext, EffectRenderContext } from '..'
 import PluginBase from './plugin-base'
-import { TypeDescriptor } from './type-descriptor'
+import { TypeDescriptor } from './TypeDescriptor'
 
 export interface EffectPluginClass {
   new (): EffectPluginBase
