+++ packages/core/src/PluginSupport/index.ts
@@ -1,7 +1,7 @@
 export { default as PluginBase } from './plugin-base'
 export { default as PostEffectBase } from './PostEffectBase'
 export { default as PluginRegistry } from './plugin-registry'
-export { default as TypeDescriptor } from './type-descriptor'
+export { default as TypeDescriptor } from './TypeDescriptor'
 
 import * as Types from './types'
 export { Types }
