+++ packages/core/src/index.ts
@@ -8,7 +8,7 @@ import { EffectRenderContext } from './Engine/RenderContext/EffectRenderContext'
 import PluginBase from './PluginSupport/plugin-base'
 import PluginRegistry from './PluginSupport/plugin-registry'
 import PostEffectBase from './PluginSupport/PostEffectBase'
-import Type, { AnyParameterTypeDescriptor, TypeDescriptor } from './PluginSupport/type-descriptor'
+import Type, { AnyParameterTypeDescriptor, TypeDescriptor } from './PluginSupport/TypeDescriptor'
 
 import * as KeyframeCalcurator from './Engine/KeyframeCalcurator'
 import * as Exporter from './Exporter'
